Rework more tests.
