--- tests/Input/Color/RenderTest.php
@@ -1,407 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPForge\Html\Tests\Input\Color;
-
-use PHPForge\Html\Input\Color;
-use PHPForge\Support\Assert;
-use PHPUnit\Framework\TestCase;
-
-/**
- * @psalm-suppress PropertyNotSetInConstructor
- */
-final class RenderTest extends TestCase
-{
-    public function testAriaDescribedBy(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" aria-describedby="MyWidget">
-            HTML,
-            Color::widget()->ariaDescribedBy('MyWidget')->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAriaLabel(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" aria-label="MyWidget">
-            HTML,
-            Color::widget()->ariaLabel('MyWidget')->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input class="class" id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->attributes(['class' => 'class'])->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testAutofocus(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" autofocus>
-            HTML,
-            Color::widget()->autofocus()->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input class="class" id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->class('class')->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testDataAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" data-test="data-value">
-            HTML,
-            Color::widget()->dataAttributes(['test' => 'data-value'])->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testDisabled(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" disabled>
-            HTML,
-            Color::widget()->disabled()->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testForm(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" form="form">
-            HTML,
-            Color::widget()->form('form')->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testGenerateAriaDescribeBy(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" aria-describedby="color-6582f2d099e8b-help">
-            HTML,
-            Color::widget()->ariaDescribedBy()->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testGenerateAriaDescribeByWithFalse(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->ariaDescribedBy(false)->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testHidden(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" hidden>
-            HTML,
-            Color::widget()->hidden()->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testId(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="id" type="color">
-            HTML,
-            Color::widget()->id('id')->render()
-        );
-    }
-
-    public function testLang(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" lang="en">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->lang('en')->render()
-        );
-    }
-
-    public function testName(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" name="name" type="color">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->name('name')->render()
-        );
-    }
-
-    public function testPrefix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            prefix
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->prefix('prefix')->render()
-        );
-    }
-
-    public function testPrefixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            prefix
-            </div>
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->prefix('prefix')->prefixContainer(true)->render()
-        );
-    }
-
-    public function testPrefixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            prefix
-            </div>
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerAttributes(['class' => 'class'])
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="class">
-            prefix
-            </div>
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerClass('class')
-                ->render()
-        );
-    }
-
-    public function testPrefixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <article>
-            prefix
-            </article>
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->prefix('prefix')
-                ->prefixContainer(true)
-                ->prefixContainerTag('article')
-                ->render()
-        );
-    }
-
-    public function testReadonly(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" readonly>
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->readonly()->render()
-        );
-    }
-
-    public function testRequired(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" required>
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->required()->render()
-        );
-    }
-
-    public function testRender(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->render()
-        );
-    }
-
-    public function testStyle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" style="style">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->style('style')->render()
-        );
-    }
-
-    public function testSuffix(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            suffix
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->suffix('suffix')->render()
-        );
-    }
-
-    public function testSuffixContainer(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            <div>
-            suffix
-            </div>
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->suffix('suffix')->suffixContainer(true)->render()
-        );
-    }
-
-    public function testSuffixContainerAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            <div class="class">
-            suffix
-            </div>
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerAttributes(['class' => 'class'])
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            <div class="class">
-            suffix
-            </div>
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerClass('class')
-                ->render()
-        );
-    }
-
-    public function testSuffixContainerTag(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            <article>
-            suffix
-            </article>
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->suffix('suffix')
-                ->suffixContainer(true)
-                ->suffixContainerTag('article')
-                ->render()
-        );
-    }
-
-    public function testTabIndex(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" tabindex="1">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->tabIndex(1)->render()
-        );
-    }
-
-    public function testTemplate(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            suffix
-            HTML,
-            Color::widget()
-                ->id('color-6582f2d099e8b')
-                ->prefix('prefix')
-                ->suffix('suffix')
-                ->template('{tag}\n{suffix}')
-                ->render()
-        );
-    }
-
-    public function testTitle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" title="title">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->title('title')->render()
-        );
-    }
-
-    public function testValue(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color" value="value">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->value('value')->render()
-        );
-    }
-
-    public function testValueWithEmpty(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <input id="color-6582f2d099e8b" type="color">
-            HTML,
-            Color::widget()->id('color-6582f2d099e8b')->value(null)->render()
-        );
-    }
-}
